kernel: use 64-bit data argument for ioctl

+++ src/fs/impl/kbdfs.cr
@@ -35,7 +35,7 @@ class KbdFSNode < VFSNode
     VFS_WAIT
   end
 
-  def ioctl(request : Int32, data : UInt32,
+  def ioctl(request : Int32, data : UInt64,
             process : Multiprocessing::Process? = nil) : Int32
     case request
     when SC_IOCTL_TCSAFLUSH
